# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Facturación, Tesorería y Previsión de Caja API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      parameters:
        - name: companyId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    patch:
      operationId: updateSupplier
      tags: [suppliers]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      parameters:
        - name: companyId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
...
  /invoices:
    get:
      operationId: listInvoices
      tags: [invoices]
      parameters:
        - name: companyId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: clientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InvoiceWithItems"
    post:
      operationId: createInvoice
      tags: [invoices]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceCreateInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceWithItems"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceWithItems"
    patch:
      operationId: updateInvoice
      tags: [invoices]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvoiceCreateInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceWithItems"
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required:
                  - success
  /invoices/{id}/status:
    patch:
      operationId: updateInvoiceStatus
      tags: [invoices]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
              required:
                - status
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceWithItems"
  /invoices/next-number:
    get:
      operationId: getNextInvoiceNumber
      tags: [invoices]
      parameters:
        - name: companyId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Next number
          content:
            application/json:
              schema:
                type: object
                properties:
                  invoiceNumber:
                    type: string
                required:
                  - invoiceNumber
...
  /invoices/{id}/payment:
    post:
      operationId: registerInvoicePayment
      tags: [invoices]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: string
                bankAccountId:
                  type: integer
                date:
                  type: string
              required:
                - amount
                - bankAccountId
      responses:
        "200":
          description: Payment registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceWithItems"
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      parameters:
        - name: companyId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskItem"
    post:
      operationId: createTask
      tags: [tasks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskItem"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskItem"
    delete:
      operationId: deleteTask
      tags: [tasks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required:
                  - success
...
components:
  schemas:
...
    Supplier:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        name:
          type: string
        taxId:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        postalCode:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        contactPerson:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        paymentTermsDays:
          type: ["integer", "null"]
        active:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, companyId, name, active, createdAt, updatedAt]
    SupplierInput:
      type: object
      properties:
        companyId:
          type: integer
        name:
          type: string
        taxId:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        postalCode:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        contactPerson:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        paymentTermsDays:
          type: integer
        active:
          type: boolean
      required: [companyId, name]
    Project:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        clientId:
          type: ["integer", "null"]
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, companyId, name, status, active, createdAt, updatedAt]
    ProjectInput:
      type: object
      properties:
        companyId:
          type: integer
        clientId:
          type: ["integer", "null"]
        name:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        active:
          type: boolean
      required: [companyId, name]
...
    InvoiceItem:
      type: object
      properties:
        id:
          type: integer
        invoiceId:
          type: integer
        description:
          type: string
        quantity:
          type: string
        unitPrice:
          type: string
        amount:
          type: string
        sortOrder:
          type: integer
      required: [id, invoiceId, description, quantity, unitPrice, amount]
    InvoiceItemInput:
      type: object
      properties:
        description:
          type: string
        quantity:
          type: string
        unitPrice:
          type: string
      required: [description, quantity, unitPrice]
    InvoiceWithItems:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        clientId:
          type: ["integer", "null"]
        projectId:
          type: ["integer", "null"]
        invoiceNumber:
          type: string
        status:
          type: string
        issueDate:
          type: string
        dueDate:
          type: ["string", "null"]
        subtotal:
          type: string
        taxRate:
          type: string
        taxAmount:
          type: string
        total:
          type: string
        paidAmount:
          type: string
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
        clientName:
          type: ["string", "null"]
        companyName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, companyId, invoiceNumber, status, issueDate, subtotal, taxRate, taxAmount, total, paidAmount, items, createdAt, updatedAt]
    InvoiceCreateInput:
      type: object
      properties:
        companyId:
          type: integer
        clientId:
          type: ["integer", "null"]
        projectId:
          type: ["integer", "null"]
        invoiceNumber:
          type: string
        status:
          type: string
        issueDate:
          type: string
        dueDate:
          type: ["string", "null"]
        taxRate:
          type: string
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItemInput"
      required: [companyId, invoiceNumber, issueDate, items]
...
    TaskItem:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: ["integer", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        priority:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        relatedType:
          type: ["string", "null"]
        relatedId:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, title, status, createdAt, updatedAt]
    TaskInput:
      type: object
      properties:
        companyId:
          type: ["integer", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
        priority:
          type: string
        dueDate:
          type: ["string", "null"]
        relatedType:
          type: ["string", "null"]
        relatedId:
          type: ["integer", "null"]
      required: [title]
...